Build improvements (#176)
* Cleaning up PanBase for easier setup and build purposes
	* Moving `PanBase` to its own file
* Proper version (i.e. `pocs.__version__` works)
