+++ setup.cfg
@@ -16,7 +16,7 @@ author_email = info@projectpanoptes.org
 description = Finding exoplanets with small digital cameras
 edit_on_github = True
 github_project = panoptes/POCS
-keywords = Citizen-science open-source exoplanet digital dslr camera
+keywords = Citizen-science open-source exoplanet digital DSLR camera astronomy STEM
 license = MIT
 long_description = PANOPTES: Panoptic Astronomical Networked Observatories for a Public Transiting Exoplanets Survey
 package_name = pocs
